Detect .cargo/config at end of any path
By the [autocmd-patterns rules][1], the current `.cargo/config` pattern
only matches `.cargo/config` as an argument.  This means that
`vim config` (from inside `.cargo`) and `vim ~/.cargo/config` do not
match.  This seems sub-optimal.  This commit adds a wildcard to the
front of the pattern so that any full path ending in `.cargo/config` is
matched.

[1]: http://vimdoc.sourceforge.net/htmldoc/autocmd.html#autocmd-patterns

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>